DT-727 Refactor tests so all relate to their own area
